Safari-611.1.10/LayoutTests/fast/xmlhttprequest/xmlhttprequest-blob-url-has-valid-response-type-and-size-expected.txt

Tests that a blob URL fetched via XHR has appropriate Content-Type and Content-Length headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.getResponseHeader('Content-Type') is "image/jpeg"
PASS xhr.getResponseHeader('Content-Length') is "764"
PASS successfullyParsed is true

TEST COMPLETE
